Fix large seeks in mk6gather
git-svn-id: https://svn.atnf.csiro.au/difx@9775 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ libraries/mark6sg/trunk/mark6gather/mark6gather.c
@@ -342,8 +342,8 @@ static void *mark6Seeker(void *arg)
 {
 	struct seekArgs *S = (struct seekArgs *)arg;
 	off_t pos;
-	int blockSize;
-	int targetBlock;
+	long long int blockSize;
+	long long int targetBlock;
 	int slotIndex;
 	int i;
 	size_t v;
